backend/backend/requirements.txt
Django>=5.2.3
djangorestframework
djangorestframework-simplejwt
python-dotenv
cloudinary
django-cloudinary-storage
sentence-transformers
numpy
django-cors-headers
PyPDF2
python-docx
xhtml2pdf
google-generativeai
psycopg2-binary
gunicorn
browser-use
